backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
fpdf2==2.7.6
email-validator==2.0.0
PyJWT==2.8.0
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
bcrypt==4.0.1
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-flask==1.3.0
coverage==7.3.2